[Chore #166103288] refactor code to solve issues in PR feedback
