+++ server/app.js
@@ -4,9 +4,8 @@ import bodyPaser from 'body-parser';
 import cors from 'cors';
 import swaggerUI from 'swagger-ui-express';
 import router from './routes/routes';
-import Db from './db/db';
-
-const swaggerDoc = require('../swagger.json');
+import Db from './db';
+import swaggerDoc from '../swagger.json';
 
 dotenv.config();
 
@@ -24,10 +23,7 @@ const PORT = process.env.PORT || 3000;
 const dbconn = new Db();
 app.listen(PORT, (err) => {
 	if (err) {
-		console.log(`error at port ${PORT}`);
-	} else {
-		// console.log(dbconn);
-		console.log(`CONNECTED at port ${PORT}`);
+		return (`error at port ${PORT}`);
 	}
 });
 export default app;
